+++ src/ui/components/button.hpp
@@ -28,7 +28,7 @@ namespace ui {
         explicit Button(
                 ServiceProvider* service_provider,
                 const std::string& text,
-                usize focus_id,
+                u32 focus_id,
                 Callback callback,
                 const Font& font,
                 const Color& text_color,
@@ -44,7 +44,7 @@ namespace ui {
                 const std::string& text,
                 const Font& font,
                 const Color& text_color,
-                usize focus_id,
+                u32 focus_id,
                 Callback callback,
                 std::pair<double, double> size,
                 Alignment alignment,
