+++ src/ui/components/slider.cpp
@@ -23,7 +23,7 @@
 
 
 ui::Slider::Slider(
-        usize focus_id,
+        u32 focus_id,
         Range range,
         Getter getter,
         Setter setter,
