+++ src/ui/components/textinput.hpp
@@ -32,7 +32,7 @@ namespace ui {
                 ServiceProvider* service_provider,
                 Font font,
                 const Color& color,
-                usize focus_id,
+                u32 focus_id,
                 const shapes::URect& fill_rect,
                 const Layout& layout,
                 bool is_top_level
@@ -43,7 +43,7 @@ namespace ui {
                 ServiceProvider* service_provider,
                 Font font,
                 const Color& color,
-                usize focus_id,
+                u32 focus_id,
                 std::pair<double, double> size,
                 Alignment alignment,
                 const Layout& layout,
